backend/requirements.txt
# FastAPI Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# HTTP & Scraping
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0

# AI/LLM
google-generativeai==0.8.3

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
python-dateutil==2.9.0
